requirements.txt
# Document Parser Requirements

# Core dependencies
python-docx>=0.8.11
Pillow>=10.0.0
openpyxl>=3.1.0

# Rules extraction
openai>=1.0.0
python-dotenv>=1.0.0

# For testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: For better testing output
colorama>=0.4.6
